% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bartlett}
\alias{bartlett}
\title{Teste de Barlett}
\usage{
bartlett(df, alpha = 0.05)
}
\arguments{
\item{df}{data.frame com dados}

\item{alpha}{número entre zero e um; nível de significancia aceito; o padrão é 0.05.}
}
\value{
\code{TRUE} se há dependência entre os dados ou
\code{FALSE} se não há (logo não faz sentido uma Análise Fatorial),
é possível retornar \code{NA} se não há linhas completas.
}
\description{
O objetivo dessa função é verificar se os dados são adequados para uma Análise Fatorial, isto é,
se há correlação entre as variáveis.
}
\section{Tipo de correlação}{


Para computar a matriz de covarância escolhemos o tipo \code{na.or.complete}, que considera apenas
as linha completas (sem \code{NA}) de \code{df}. Se não houver linhas completas, o valor retornado será
um \code{NA}. Originalmente usava-se o \code{pairwise.complete.obs}, mas isso pode retornar matrizes de
correlçao com autoravalores negativos, o que não é adequado.
Esses métodos de computação da matriz de covariância estão disponíveis no argumento
\code{use} da função \code{stats::cor}.
}

