% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_sim.R
\name{base_sim}
\alias{base_sim}
\title{Base simulation function}
\usage{
base_sim(
  dataset,
  close_matrix,
  sim_max,
  bhv_increment,
  CT_attribution_fraction,
  WT_attribution_fraction
)
}
\arguments{
\item{dataset}{A dataset set up so that variable 1 is ID, variable 2 is behavior at time 0 (0-1 scale by fixed intervals), variable 3 is the close ties response, and variable 4 is the weak ties response. Total of 4 variables.}

\item{close_matrix}{A unidirectional or bidirectional adjacency matrix. Values are either 0 or 1 to represent connections.}

\item{sim_max}{A value at which to limit the simulation. Warning, it may never reach that value if values do not improve from t to t+1.}

\item{bhv_increment}{The increments at which your behavior variable increases. It is a constant value.}

\item{CT_attribution_fraction}{The weight of close ties.}

\item{WT_attribution_fraction}{The weight of weak ties.}
}
\value{
Returns a dataset, labeled ndata, to the global environment with simulation results.
}
\description{
Base simulation function
}
