% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dados_snis}
\alias{dados_snis}
\title{Dados}
\format{
An object of class \code{list} of length 18.
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
dados_snis
}
\description{
Dados formatados
}
\keyword{datasets}
