% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa.R
\name{fa}
\alias{fa}
\title{Análise Fatorial para dados do SNIS}
\usage{
fa(df, grupos, normalizar = T, rotacao = "oblimin")
}
\arguments{
\item{df}{data.frame com dados para análise fatorial}

\item{grupos}{data.frame com esquema de grupos para a análise fatorial, pode ser passada string 'eee' ou 'su'}

\item{normalizar}{lógico, \code{TRUE} (padrão) se os scores devem ser padronizados, \code{FALSE} se não.}

\item{rotacao}{string, tipo de rotação a ser usada no argumento \code{rotate} da função \code{psych::fa()}, o padrão é 'oblimin'}
}
\value{
objeto do tipo \code{fa-snis}, uma lista contendo os scores e loadings de cada grupo
}
\description{
Função geradora dos scores e loadings da análise fatorial com dados do SINS
}
\section{Grupos}{


O argumento grupos deve ser um data.frame contendo as colunas.
}

