% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapa_MG}
\alias{mapa_MG}
\title{Mapa de Minas Gerais como json}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 853 rows and 4 columns.
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
mapa_MG
}
\description{
Dados formatados
}
\keyword{datasets}
