% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_sim.R
\name{Order_of_influence}
\alias{Order_of_influence}
\title{Most influential individuals}
\usage{
Order_of_influence(ind_data_output, testing_time_avg)
}
\arguments{
\item{ind_data_output}{Requires output dataset from ind_approach function.}

\item{testing_time_avg}{The average value from the final time period of the tt function. Used to standardize results.}
}
\value{
Returns a dataset, labeled influential_inds, to the global environment with simulation results.
}
\description{
Generates a dataset of the most influential individuals from the individual approach(ind_approach function).
}
