% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Title}
\usage{
plot_boxplot(df, score_col, group_col, titulo = NULL, cor_paleta = "Set2")
}
\arguments{
\item{df}{.}

\item{score_col}{.}

\item{group_col}{.}

\item{titulo}{.}

\item{cor_paleta}{.}
}
\value{
plot
}
\description{
Title
}
