% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_density_score}
\alias{plot_density_score}
\title{Densidade de uma variável numérica}
\usage{
plot_density_score(df, group_col, titulo = NULL)
}
\arguments{
\item{df}{Data frame contendo os dados.}

\item{titulo}{(opcional) Título do gráfico. Se \code{NULL}, é gerado automaticamente.}

\item{var_col}{String com o nome da coluna numérica a ser analisada, como \code{"IN023"}.}
}
\value{
Um objeto \code{ggplot} com o gráfico de densidade.
}
\description{
Cria um gráfico de densidade mostrando a distribuição suave dos valores
de uma variável numérica, como uma "onda".
}
