% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_hist_score}
\alias{plot_hist_score}
\title{Histograma de uma variável numérica}
\usage{
plot_hist_score(df, group_col, titulo = NULL)
}
\arguments{
\item{df}{Data frame contendo os dados.}

\item{titulo}{(opcional) Título do gráfico. Se \code{NULL}, é gerado automaticamente.}

\item{var_col}{String com o nome da coluna numérica a ser analisada, como \code{"IN023"}.}
}
\value{
Um objeto \code{ggplot} com o histograma.
}
\description{
Cria um histograma mostrando a distribuição dos valores de uma variável numérica.
}
