% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quartis}
\alias{quartis}
\title{Criar vetor de quartis}
\usage{
quartis(vec)
}
\arguments{
\item{vec}{vetor de números a serem analisados}
}
\value{
vetor com a faixa quartílica de cada número
}
\description{
Cria um vetor que mostra em que faixa dos quartis os elementos do vetor original estão
}
