% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_sim.R
\name{riskiest_ppl}
\alias{riskiest_ppl}
\title{Riskiest people dataset}
\usage{
riskiest_ppl(tt_output, max_bhv)
}
\arguments{
\item{tt_output}{Requires output dataset from tt function.}

\item{max_bhv}{Maximum behavior extraction value.}
}
\value{
Returns a dataset, labeled riskiest_ppl, to the global environment with simulation results from tt.
}
\description{
Generates a dataset of individuals(and associated results) from testing time simulation.
These are the riskiest individuals at the end of tt simulation -- Time does not fix these individuals!
}
