% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_sim.R
\name{sim_dataset}
\alias{sim_dataset}
\title{Base simulation dataset generator}
\usage{
sim_dataset(dataset, close_matrix)
}
\arguments{
\item{dataset}{A dataset set up so that variable 1 is ID, variable 2 is behavior at time 0 (0-1 scale by fixed intervals), variable 3 is the close ties response, and variable 4 is the weak ties response. Total of 4 variables.}

\item{close_matrix}{A unidirectional or bidirectional adjacency matrix. Values are either 0 or 1 to represent connections.}
}
\value{
Returns a dataset, labeled ndata, to the global environment with simulation results.
}
\description{
Base simulation dataset generator
}
