% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{table_median}
\alias{table_median}
\title{Title}
\usage{
table_median(df, score_col = "score médio eee", group_col = "Nome_Mesorregião")
}
\arguments{
\item{df}{.}

\item{score_col}{.}

\item{group_col}{.}
}
\value{
plot
}
\description{
Title
}
