% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{table_top_bottom}
\alias{table_top_bottom}
\title{Title}
\usage{
table_top_bottom(df, top_n = 10, bottom_n = 10)
}
\arguments{
\item{df}{.}

\item{top_n}{.}

\item{bottom_n}{.}
}
\value{
table
}
\description{
Title
}
