% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_df_fa}
\alias{update_df_fa}
\title{Atualizar data frame com scores}
\usage{
update_df_fa(df, ...)
}
\arguments{
\item{df}{data.frame usado para fazer análise fatorial}

\item{...}{objetos do tipo \code{fa-snis}}
}
\value{
data.frame atualizado
}
\description{
Adiciona colunas no data frame original com os valores dos scores desejados
}
