% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_sim.R
\name{weak_matrix}
\alias{weak_matrix}
\title{Weak-Ties matrix generator}
\usage{
weak_matrix(ndata)
}
\arguments{
\item{ndata}{A dataset set up so that variable 1 is ID, variable 2 is behavior at time 0 (0-1 scale by fixed intervals), variable 3 is the close ties response, and variable 4 is the weak ties response. Total of 4 variables.}
}
\value{
Returns to the global environment a weak-ties adjacency matrix labeled adj_weak.
}
\description{
Generates an adjacency matrix for an environment where everyone is connected to everyone but themselves. Attempts to capture the social network at large.
}
