% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Boxplots Scores por Grupos}
\usage{
plot_boxplot(df, score_cols, group_cols, titulo = NULL, cor_paleta = "plasma")
}
\arguments{
\item{df}{Data frame contendo os dados.}

\item{score_cols}{Vetor de strings com os nomes das colunas que contêm os scores, como \verb{score médio su}, \verb{score médio eee}.}

\item{group_cols}{Vetor de strings com os nomes das colunas de agrupamento, como \verb{natureza jurídica}.}

\item{titulo}{(opcional) Título do gráfico. Se \code{NULL}, é gerado automaticamente}

\item{cor_paleta}{(opcional) Paleta de cores viridis a ser usada.
Opções: \code{"viridis"}, \code{"plasma"}, \code{"magma"}, \code{"cividis"}, \code{"inferno"}.
Padrão: \code{"plasma"}.}
}
\value{
Um objeto \code{ggplot} contendo o boxplot formatado.
}
\description{
Essa função gera um boxplot que mostra a distribuição de uma variável de
\emph{score} em diferentes grupos (como prestadores, regiões ou naturezas jurídicas).
}
