% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_interactive_map}
\alias{plot_interactive_map}
\title{Title}
\usage{
plot_interactive_map(df, score_col, group_col, titulo = NULL)
}
\arguments{
\item{df}{.}

\item{score_col}{.}

\item{group_col}{.}

\item{titulo}{.}
}
\value{
tmap
}
\description{
Title
}
