% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_median_barplot}
\alias{plot_median_barplot}
\title{Title}
\usage{
plot_median_barplot(
  df,
  score_col,
  group_col,
  titulo = NULL,
  casas_decimais = 3,
  cor_paleta = "Blues"
)
}
\arguments{
\item{df}{.}

\item{score_col}{.}

\item{group_col}{.}

\item{titulo}{.}

\item{casas_decimais}{.}

\item{cor_paleta}{.}
}
\value{
plot
}
\description{
Title
}
