% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_ridge_scores}
\alias{plot_ridge_scores}
\title{Title}
\usage{
plot_ridge_scores(df, group_col, score_cols, titulo = NULL)
}
\arguments{
\item{df}{.}

\item{group_col}{.}

\item{score_cols}{.}

\item{titulo}{.}
}
\value{
plot
}
\description{
Title
}
