% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processar_dados.R
\name{processar_dados}
\alias{processar_dados}
\title{Processar dados para análise fatorial}
\usage{
processar_dados(ano)
}
\arguments{
\item{ano}{inteiro entre 2000 e 2022}
}
\value{
tibble
}
\description{
Rotina de processamento (e criação) de variáveis para a aplicação de análise fatorial
}
\section{Inputação}{


A inputação foi feita pelo pacote \code{mice} em duas etapas. Detectamos que as seguintes variáveis têm
forte correlação entre si: IN024 com IN047; AG022 com AG013. Como essa forte multicolinearidade piora
a qualidade das previsões feitas pelo \code{mice}, decidimos excluindo IN047 e AG013 para a primeira
fase de imputação (em que IN024 e AG022 são imputadas). Em seguida, usamos os valores imputados
para prever IN047 e AG013.

O pacote \code{mice} usa alguma aleatoriedade para suas previsões, por isso decidimos fixar a semente
das previsões para que os mesmos resultados sejam retornados em cada ano.

Os valores de algumas variáveis foram imputadas por simples lógica, como as que se seguem:
\itemize{
\item \code{IN003}: ;
}
}

\section{Variáveis criadas}{


Algumas variáveis foram criadas com as seguintes fórmulas:
\itemize{
\item \code{tarifa}: ;
\item \code{micromedida}: ;
\item \code{urbanização}: ;
\item \code{prestador2}: ;
}
}

